# indexer/openapi.yaml
openapi: 3.1.0
info:
  title: Frontier Orderbook Indexer API
  version: 0.1.0
  description: >
    REST API over the indexed Frontier on-chain orderbook. The indexer watches
    book / factory / position-NFT events and serves normalized markets,
    positions, fills/trades, account rollups, and depth snapshots from SQLite.
    All token amounts are returned as decimal strings of the underlying uint256
    so no precision is lost over JSON.
servers:
  - url: http://localhost:8787
    description: Local indexer
paths:
  /health:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  ts: { type: integer }
  /markets:
    get:
      summary: List indexed markets
      responses:
        "200":
          description: All known books
          content:
            application/json:
              schema:
                type: object
                properties:
                  markets:
                    type: array
                    items: { $ref: "#/components/schemas/Market" }
  /book/{market}:
    get:
      summary: Depth snapshot for a market
      description: >
        Aggregated live bid/ask depth reconstructed from indexed positions.
        Approximates FrontierLens.depth for the uniform/geometric book.
      parameters:
        - { $ref: "#/components/parameters/Market" }
        - name: levels
          in: query
          required: false
          schema: { type: integer, default: 200, maximum: 1000 }
      responses:
        "200":
          description: Depth snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Depth" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /positions/{owner}:
    get:
      summary: Positions for an owner
      parameters:
        - { $ref: "#/components/parameters/Owner" }
        - name: live
          in: query
          required: false
          description: When "true", return only live positions.
          schema: { type: string, enum: ["true", "false"] }
      responses:
        "200":
          description: Owner positions
          content:
            application/json:
              schema:
                type: object
                properties:
                  owner: { type: string }
                  positions:
                    type: array
                    items: { $ref: "#/components/schemas/Position" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /trades:
    get:
      summary: Filterable taker trade tape
      parameters:
        - name: market
          in: query
          schema: { type: string }
        - name: taker
          in: query
          schema: { type: string }
        - name: side
          in: query
          schema: { type: string, enum: [buy, sell] }
        - name: fromBlock
          in: query
          schema: { type: integer }
        - name: toBlock
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
        - name: cursor
          in: query
          description: >
            Opaque keyset cursor from a prior response's `nextCursor`. Continues
            strictly after that point in the (block, logIndex) ordering so
            pagination is stable as new trades are appended.
          schema: { type: string }
      responses:
        "200":
          description: Matching trades, newest first, with a pagination cursor
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items: { $ref: "#/components/schemas/Trade" }
                  nextCursor:
                    type: [string, "null"]
                    description: Pass back as `cursor` for the next page; null at end.
        "400": { $ref: "#/components/responses/BadRequest" }
  /stats/{market}:
    get:
      summary: 24h-style stats for a market
      description: >
        Trade count, taker volume (per input token), token1 fill volume, taker
        fees (per token), and a point-in-time open-interest snapshot over a
        sliding window ending at the current time.
      parameters:
        - { $ref: "#/components/parameters/Market" }
        - name: window
          in: query
          required: false
          description: Window size in seconds (default 86400 = 24h).
          schema: { type: integer, default: 86400, minimum: 1 }
      responses:
        "200":
          description: Market stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketStats" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /candles/{market}:
    get:
      summary: OHLC candles derived from interval fills
      description: >
        Open/high/low/close price (1.0001^tick) plus token1 volume, bucketed by
        block timestamp into `interval`-second windows. For charting.
      parameters:
        - { $ref: "#/components/parameters/Market" }
        - name: interval
          in: query
          required: false
          description: Bucket size in seconds (e.g. 60, 300, 3600).
          schema: { type: integer, default: 3600, minimum: 1 }
        - name: from
          in: query
          schema: { type: integer, description: "unix seconds lower bound" }
        - name: to
          in: query
          schema: { type: integer, description: "unix seconds upper bound" }
        - name: limit
          in: query
          schema: { type: integer, default: 1000, maximum: 5000 }
      responses:
        "200":
          description: OHLC candles, oldest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  market: { type: string }
                  interval: { type: integer }
                  candles:
                    type: array
                    items: { $ref: "#/components/schemas/Candle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /account/{owner}:
    get:
      summary: Per-account rollup across markets, incl. claim tokens
      parameters:
        - { $ref: "#/components/parameters/Owner" }
      responses:
        "200":
          description: Account summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  parameters:
    Market:
      name: market
      in: path
      required: true
      schema: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
    Owner:
      name: owner
      in: path
      required: true
      schema: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
  responses:
    BadRequest:
      description: Invalid parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Market:
      type: object
      properties:
        address: { type: string }
        token0: { type: string }
        token1: { type: string }
        tickSpacing: { type: integer }
        startTick: { type: [integer, "null"] }
        currentTick: { type: [integer, "null"] }
        makerFeeBps: { type: integer }
        takerFeeBps: { type: integer }
        feeRecipient: { type: [string, "null"] }
        hooks: { type: [string, "null"] }
        createdBlock: { type: [integer, "null"] }
    DepthLevel:
      type: object
      properties:
        tick: { type: integer }
        askSize: { type: string, description: "token0 resting (decimal string)" }
        bidSize: { type: string, description: "token0-denominated bid (decimal string)" }
    Depth:
      type: object
      properties:
        market: { type: string }
        currentTick: { type: [integer, "null"] }
        bids:
          type: array
          items: { $ref: "#/components/schemas/DepthLevel" }
        asks:
          type: array
          items: { $ref: "#/components/schemas/DepthLevel" }
    ClaimToken:
      type: object
      description: ERC-721 token (FrontierPositionNFT) wrapping a book position.
      properties:
        wrapper: { type: string }
        tokenId: { type: string }
        owner: { type: string }
    Position:
      type: object
      properties:
        market: { type: string }
        positionId: { type: string }
        owner: { type: string }
        lowerTick: { type: integer }
        upperTick: { type: integer }
        liquidity: { type: string }
        isBid: { type: boolean }
        live: { type: boolean }
        claimed: { type: string, description: "cumulative proceeds1 claimed" }
        depositBlock: { type: [integer, "null"] }
        claimToken:
          oneOf:
            - { $ref: "#/components/schemas/ClaimToken" }
            - { type: "null" }
    Trade:
      type: object
      properties:
        market: { type: string }
        taker: { type: string }
        token: { type: string, description: "fee/input token" }
        grossInput: { type: string }
        fee: { type: string }
        totalPaid: { type: string }
        side: { type: [string, "null"], enum: [buy, sell, null] }
        blockNumber: { type: integer }
        logIndex: { type: integer }
        txHash: { type: string }
        timestamp: { type: [integer, "null"] }
    MarketStats:
      type: object
      properties:
        market: { type: string }
        windowSecs: { type: integer }
        since: { type: [integer, "null"] }
        tradeCount: { type: integer }
        volumeByToken:
          type: object
          additionalProperties: { type: string }
          description: Taker gross input volume per input token (decimal strings).
        volume1: { type: string, description: "token1 fill volume (sum proceeds1)" }
        feesByToken:
          type: object
          additionalProperties: { type: string }
        openInterest: { type: string, description: "token0 resting liquidity * levels" }
        livePositions: { type: integer }
    Candle:
      type: object
      properties:
        time: { type: integer, description: "bucket start, unix seconds" }
        open: { type: number }
        high: { type: number }
        low: { type: number }
        close: { type: number }
        volume: { type: string, description: "token1 volume in the bucket" }
        trades: { type: integer }
    Account:
      type: object
      properties:
        owner: { type: string }
        markets:
          type: array
          items:
            type: object
            properties:
              market: { type: string }
              livePositions: { type: integer }
              totalPositions: { type: integer }
              proceedsClaimed: { type: string }
              principalReturned: { type: string }
              feesPaidMaker: { type: string }
              feesPaidTaker: { type: string }
              updatedBlock: { type: [integer, "null"] }
        claimTokens:
          type: array
          items:
            type: object
            properties:
              wrapper: { type: string }
              tokenId: { type: string }
              market: { type: [string, "null"] }
              positionId: { type: [string, "null"] }
  # WebSocket channels (documented here; not part of the OpenAPI HTTP surface):
  #   ws://localhost:8787/fills?market=0x...   -> {type:"fills", data:{...}}
  #   ws://localhost:8787/depth?market=0x...   -> {type:"depth", data:{...}}